Primitive crate
