+++ Cargo.toml
@@ -122,6 +122,7 @@ members = [
 	"primitives/consensus/pow",
 	"primitives/consensus/vrf",
 	"primitives/core",
+	"primitives/chain-spec",
 	"primitives/debug-derive",
 	"primitives/storage",
 	"primitives/externalities",
