+++ client/chain-spec/Cargo.toml
@@ -16,6 +16,7 @@ sp-core = { version = "2.0.0-alpha.5", path = "../../primitives/core" }
 serde = { version = "1.0.101", features = ["derive"] }
 serde_json = "1.0.41"
 sp-runtime = { version = "2.0.0-alpha.5", path = "../../primitives/runtime" }
+sp-chain-spec = { version = "2.0.0-alpha.5", path = "../../primitives/chain-spec" }
 sc-telemetry = { version = "2.0.0-alpha.5", path = "../telemetry" }
 
 [package.metadata.docs.rs]
