+++ client/rpc-api/Cargo.toml
@@ -21,7 +21,7 @@ parking_lot = "0.10.0"
 sp-core = { version = "2.0.0-alpha.5", path = "../../primitives/core" }
 sp-version = { version = "2.0.0-alpha.5", path = "../../primitives/version" }
 sp-runtime = { path = "../../primitives/runtime" , version = "2.0.0-alpha.5"}
-sc-chain-spec = { path = "../chain-spec" , version = "2.0.0-alpha.5"}
+sp-chain-spec = { path = "../../primitives/chain-spec" , version = "2.0.0-alpha.5"}
 serde = { version = "1.0.101", features = ["derive"] }
 serde_json = "1.0.41"
 sp-transaction-pool = { version = "2.0.0-alpha.5", path = "../../primitives/transaction-pool" }
