+++ client/rpc-api/src/system/mod.rs
@@ -27,7 +27,7 @@ use self::error::Result as SystemResult;
 
 pub use self::helpers::{SystemInfo, Health, PeerInfo, NodeRole};
 pub use self::gen_client::Client as SystemClient;
-use sc_chain_spec::Properties;
+use sp_chain_spec::Properties;
 
 /// Substrate system RPC API
 #[rpc]
@@ -46,7 +46,7 @@ pub trait SystemApi<Hash, Number> {
 
 	/// Get the chain's type.
 	#[rpc(name = "system_chainType")]
-	fn system_type(&self) -> SystemResult<sc_chain_spec::ChainType>;
+	fn system_type(&self) -> SystemResult<sp_chain_spec::ChainType>;
 
 	/// Get a custom set of properties as a JSON object, defined in the chain spec.
 	#[rpc(name = "system_properties")]
